+++ src/Homes/Content.js
@@ -3,13 +3,13 @@ import styled from "styled-components";
 import Card from "./Card";
 import Pagination from "./Pagination";
 
-import home1 from "./home-1.png";
-import home2 from "./home-2.png";
-import home3 from "./home-3.png";
-import home4 from "./home-4.png";
-import home5 from "./home-5.png";
-import home6 from "./home-6.png";
-import pin from "./pin.svg";
+import home1 from "./UI/home-1.png";
+import home2 from "./UI/home-2.png";
+import home3 from "./UI/home-3.png";
+import home4 from "./UI/home-4.png";
+import home5 from "./UI/home-5.png";
+import home6 from "./UI/home-6.png";
+import pin from "./UI/pin.svg";
 
 const Content = styled.div`
   margin-top: 24px;
