--- src/Homes/react_dates_overrides.css
@@ -1,36 +1,0 @@
-.DayPicker__horizontal {
-  box-shadow: none !important;
-  margin: 0;
-}
-.DayPicker__vertical {
-  padding: 0;
-}
-
-.DayPickerKeyboardShortcuts_show__bottomRight,
-.DayPickerKeyboardShortcuts_show__topRight {
-  display: none;
-}
-.CalendarMonth_table {
-  margin-top: 10px;
-}
-.CalendarDay_container {
-  border: 1px solid #d5d5d5 !important;
-  color: #383838;
-  font-size: 14px;
-  line-height: 16px;
-}
-.DayPicker_weekHeader_li {
-  font-size: 14px;
-  line-height: 16px;
-  color: #636363;
-}
-.CalendarDay__blocked_calendar .CalendarDay_button {
-  color: #cbcbcb;
-  background-color: #fff;
-}
-.Calendarmonth_caption .Calendarmonth_caption_1 {
-  font-size: 18px;
-  line-height: 21px;
-  font-family: CircularBold, "Helvetica Neue", Helvetica, sans-serif;
-  color: #383838;
-}
